+++ src/components/functions.ts
@@ -4,22 +4,49 @@ import type { IntlShape } from "react-intl";
 import { getArmors, getMaterials } from "../data";
 import { type IArmor } from "../redux/armors";
 
+type IItemCounts = [string, number];
+export interface IItemsCounts {
+  items: IItemCounts[];
+  sum: number;
+}
 interface IItems {
-  all: number;
-  selection: number;
+  all: IItemsCounts;
+  selection: IItemsCounts;
 }
 export interface ICalculatedItems extends IItems {
   id: string;
   name: string;
+  slug: string;
 }
+
+export const deduplicateItems = (items: IItemCounts[]): IItemsCounts => {
+  const deduplicatedItems: Record<string, number> = {};
+  let totalSum = 0;
+
+  for (const [item, value] of items) {
+    if (deduplicatedItems[item] !== undefined) {
+      (deduplicatedItems[item] as number) += value;
+    } else {
+      deduplicatedItems[item] = value;
+    }
+  }
+
+  const deduplicatedItemsArray: [string, number][] = [];
+  for (const item in deduplicatedItems) {
+    deduplicatedItemsArray.push([item, deduplicatedItems[item] as number]);
+    totalSum += deduplicatedItems[item] as number;
+  }
+
+  return { items: deduplicatedItemsArray.sort((a, b) => a[0].localeCompare(b[0])), sum: totalSum };
+};
 
 export const calculateListItems = (
   armorsState: IArmor[],
   intl: IntlShape
 ): { items: ICalculatedItems[]; rupee: IItems; rupeeBuy: IItems } => {
-  const items: Record<string, { all: number; name: string; selection: number }> = {};
-  const rupee: IItems = { all: 0, selection: 0 };
-  const rupeeBuy: IItems = { all: 0, selection: 0 };
+  const items: Record<string, { all: IItemCounts[]; name: string; selection: IItemCounts[] }> = {};
+  const rupeeCount: { all: IItemCounts[]; selection: IItemCounts[] } = { all: [], selection: [] };
+  const rupeeBuyCount: { all: IItemCounts[]; selection: IItemCounts[] } = { all: [], selection: [] };
 
   const armors = getArmors(intl);
   const materials = getMaterials(intl);
@@ -33,31 +60,43 @@ export const calculateListItems = (
       const getStatus = getActiveStatus(armor);
       const getStatusAll = getActiveStatus({ ...armor, wanted: 4 });
 
+      if (!armor.owned && armorData.buyPrice !== undefined) {
+        rupeeBuyCount.all.push([armorData.name, armorData.buyPrice]);
+      }
+
       const status1 = getStatus(1);
       if (status1) {
         materialsSelection.push(...armorData.rank1);
-        rupee.selection += 10;
+
+        rupeeCount.selection.push([armorData.name, 10]);
+
+        if (!armor.owned && armorData.buyPrice !== undefined) {
+          rupeeBuyCount.selection.push([armorData.name, armorData.buyPrice]);
+        }
       }
 
       const status1All = getStatusAll(1);
 
       if (status1All) {
         materialsAll.push(...armorData.rank1);
-        rupee.all += 10;
+
+        rupeeCount.all.push([armorData.name, 10]);
       }
 
       if (armorData.rank2) {
         const status2 = getStatus(2);
         if (status2) {
           materialsSelection.push(...armorData.rank2);
-          rupee.selection += 50;
+
+          rupeeCount.selection.push([armorData.name, 50]);
         }
 
         const status2All = getStatusAll(2);
 
         if (status2All) {
           materialsAll.push(...armorData.rank2);
-          rupee.all += 50;
+
+          rupeeCount.all.push([armorData.name, 50]);
         }
       }
 
@@ -65,14 +104,16 @@ export const calculateListItems = (
         const status3 = getStatus(3);
         if (status3) {
           materialsSelection.push(...armorData.rank3);
-          rupee.selection += 200;
+
+          rupeeCount.selection.push([armorData.name, 200]);
         }
 
         const status3All = getStatusAll(3);
 
         if (status3All) {
           materialsAll.push(...armorData.rank3);
-          rupee.all += 200;
+
+          rupeeCount.all.push([armorData.name, 200]);
         }
       }
 
@@ -80,14 +121,16 @@ export const calculateListItems = (
         const status4 = getStatus(4);
         if (status4) {
           materialsSelection.push(...armorData.rank4);
-          rupee.selection += 500;
+
+          rupeeCount.selection.push([armorData.name, 500]);
         }
 
         const status4All = getStatusAll(4);
 
         if (status4All) {
           materialsAll.push(...armorData.rank4);
-          rupee.all += 500;
+
+          rupeeCount.all.push([armorData.name, 500]);
         }
       }
 
@@ -95,34 +138,49 @@ export const calculateListItems = (
         const name = materials.find((material) => material.id === id)?.name ?? id;
 
         if (!items[name]) {
-          items[name] = { all: 0, name, selection: 0 };
+          items[name] = { all: [], name, selection: [] };
         }
 
-        items[name].selection += count;
+        items[name].selection.push([armorData.name, count]);
       });
 
       materialsAll.forEach(([id, count]) => {
         const name = materials.find((material) => material.id === id)?.name ?? id;
         if (!items[name]) {
-          items[name] = { all: 0, name, selection: 0 };
+          items[name] = { all: [], name, selection: [] };
         }
-
-        items[name].all += count;
+        items[name].all.push([armorData.name, count]);
       });
     }
   });
 
-  const array: ICalculatedItems[] = Object.keys(items).map((id) => ({
-    all: items[id].all ?? 0,
-    id: id,
-    name: id,
-    selection: items[id].selection ?? 0
-  }));
+  const array: ICalculatedItems[] = Object.keys(items).map((id) => {
+    /*const armors = Array.from(
+      new Set([...items[id].all.map(([armor]) => armor), ...items[id].selection.map(([armor]) => armor)])
+    );*/
+
+    const all: IItemsCounts = deduplicateItems(items[id].all);
+    const selection: IItemsCounts = deduplicateItems(items[id].selection);
+
+    return {
+      all,
+      id: id,
+      name: id,
+      selection,
+      slug: getSlug(id)
+    };
+  });
+
+  const rupeeAll = deduplicateItems(rupeeCount.all);
+  const rupeeSelection = deduplicateItems(rupeeCount.selection);
+
+  const rupeeBuyAll = deduplicateItems(rupeeBuyCount.all);
+  const rupeeBuySelection = deduplicateItems(rupeeBuyCount.selection);
 
   return {
     items: array.sort((a, b) => a.name.localeCompare(b.name)),
-    rupee,
-    rupeeBuy
+    rupee: { all: rupeeAll, selection: rupeeSelection },
+    rupeeBuy: { all: rupeeBuyAll, selection: rupeeBuySelection }
   };
 };
 
@@ -182,3 +240,5 @@ export const useMediaQuery = (query: string): boolean => {
 
   return matches;
 };
+
+export const getSlug = (name: string) => name.toLowerCase().replace(/[^a-z]/g, "");